Move consts back to const
